% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b.complexICA.R
\name{complex_ICA_modules}
\alias{complex_ICA_modules}
\title{Uses ICA to find the core of modules, then iteratively expands outward based on a decreasing threshold of correlation}
\usage{
complex_ICA_modules(
  x,
  n.comp,
  core.size = 5,
  starting.threshold = 0.5,
  min.threshold = 0
)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{n.comp}{the number of components to search for in the data}

\item{core.size}{the size of the module cores determine by ICA}

\item{starting.threshold}{the starting threshold for correlation to be added to a module}

\item{min.threshold}{the minimun threshold for correlation, if set above zero will result in incomplete assignments}
}
\value{
a module object
}
\description{
Uses ICA to find the core of modules, then iteratively expands outward based on a decreasing threshold of correlation
}
