% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outdated.R
\name{.label_mods}
\alias{.label_mods}
\title{Helper to label the modules of a graph generated by make_modular_graph}
\usage{
.label_mods(g, n.nodes, n.mods, mod.names)
}
\arguments{
\item{g}{an igraph object whose nodes are to be labeled with module membership}

\item{n.nodes}{an integer, the total number of nodes in the graph}

\item{n.mods}{an integer, the number of modules in the graph}

\item{mod.names}{a character vector of length n.mods, the module label for each module}
}
\value{
an igraph object with labeled modules for each node
}
\description{
Helper to label the modules of a graph generated by make_modular_graph
}
\keyword{internal}
