% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.lower_cut_height}
\alias{.lower_cut_height}
\title{Helper to find_WGCNA_mods that lowers the dynamic-tree-cut height until a
minimum number of modules is produced}
\usage{
.lower_cut_height(
  index.vector,
  dendro,
  dis,
  min.mods,
  min.size,
  cut.height = NULL,
  max.size = Inf,
  min.cut.height = 0
)
}
\arguments{
\item{index.vector}{an integer vector of length p, the modules from the initial cut (0 = unassigned)}

\item{dendro}{a dendrogram (flashClust/hclust object) used for the cut}

\item{dis}{a p x p TOM distance matrix passed to cutreeDynamic}

\item{min.mods}{an integer, the minimum number of modules required}

\item{min.size}{an integer, the minimum number of nodes allowed in a module}

\item{cut.height}{a numeric or NULL, the starting cut height; recomputed from the dendrogram when NULL}

\item{max.size}{a numeric, the maximum module size (used only for messaging)}

\item{min.cut.height}{a numeric, the floor below which the cut height will not be lowered}
}
\value{
a list with the updated index.vector and the cut.height that produced it
}
\description{
Helper to find_WGCNA_mods that lowers the dynamic-tree-cut height until a
minimum number of modules is produced
}
\keyword{internal}
