% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_WGCNA_mods.R
\name{.merge_small_modules}
\alias{.merge_small_modules}
\title{Helper to find_WGCNA_mods that merges modules below min.size into their
nearest neighbouring module (by mean TOM distance) when there is room}
\usage{
.merge_small_modules(index.vector, dis, min.size, max.size)
}
\arguments{
\item{index.vector}{an integer vector of length p assigning each node to a module (0 = unassigned)}

\item{dis}{a p x p TOM distance matrix}

\item{min.size}{an integer, the minimum number of nodes allowed in a module}

\item{max.size}{a numeric, the maximum module size a merge target may reach}
}
\value{
an integer vector of length p with undersized modules merged where possible
}
\description{
Helper to find_WGCNA_mods that merges modules below min.size into their
nearest neighbouring module (by mean TOM distance) when there is room
}
\keyword{internal}
