% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_WGCNA_mods.R
\name{.n_modules}
\alias{.n_modules}
\title{Count the modules in an index vector, excluding the unassigned group (0)}
\usage{
.n_modules(index.vector)
}
\arguments{
\item{index.vector}{an integer vector assigning each node to a module (0 = unassigned)}
}
\value{
an integer, the number of distinct non-zero modules
}
\description{
Count the modules in an index vector, excluding the unassigned group (0)
}
\keyword{internal}
