% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sft_check.R
\name{.sft_check}
\alias{.sft_check}
\title{Select soft-thresholding power from WGCNA::pickSoftThreshold output}
\usage{
.sft_check(sft)
}
\arguments{
\item{sft}{a list, the output of WGCNA::pickSoftThreshold}
}
\value{
an integer, the selected soft-thresholding power
}
\description{
Select soft-thresholding power from WGCNA::pickSoftThreshold output
}
\keyword{internal}
