% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.split_large_modules}
\alias{.split_large_modules}
\title{Helper to find_WGCNA_mods that splits modules exceeding max.size by
re-clustering each oversized module}
\usage{
.split_large_modules(
  index.vector,
  dis,
  max.size,
  min.size,
  hclust.method = "average",
  cut.height = NULL
)
}
\arguments{
\item{index.vector}{an integer vector of length p assigning each node to a module (0 = unassigned)}

\item{dis}{a p x p TOM distance matrix}

\item{max.size}{a numeric, the maximum number of nodes allowed in a module}

\item{min.size}{an integer, the minimum number of nodes allowed in a module}

\item{hclust.method}{a character, the agglomeration method passed to flashClust}

\item{cut.height}{a numeric or NULL, the cut height passed to cutreeDynamic}
}
\value{
an integer vector of length p with oversized modules split where possible
}
\description{
Helper to find_WGCNA_mods that splits modules exceeding max.size by
re-clustering each oversized module
}
\keyword{internal}
