% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_SILGGM_graph.R
\name{.upper_tri_to_matrix}
\alias{.upper_tri_to_matrix}
\title{Reconstruct the symmetric matrix from upper triangular vector}
\usage{
.upper_tri_to_matrix(upper_tri_values, variable_names = NULL, diagl = 1)
}
\arguments{
\item{upper_tri_values}{a numeric vector of the upper triangle of the matrix}

\item{variable_names}{a character vector of row and column names for the output matrix}

\item{diagl}{a numeric scalar or vector of length p, the diagonal value(s) of the output matrix}
}
\value{
a symmetric matrix
}
\description{
Reconstruct the symmetric matrix from upper triangular vector
}
\keyword{internal}
