% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_SILGGM_graph.R
\name{.upper_tri_vec}
\alias{.upper_tri_vec}
\title{Take the upper triangular vector of a symmetric matrix}
\usage{
.upper_tri_vec(mat)
}
\arguments{
\item{mat}{the input symmetric matrix}
}
\value{
a numeric vector
}
\description{
Take the upper triangular vector of a symmetric matrix
}
\keyword{internal}
