% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outdated.R
\name{find_mcl_mods}
\alias{find_mcl_mods}
\title{Detect co-expression modules from a data matrix using Markov Clustering (MCL)}
\usage{
find_mcl_mods(
  x,
  beta = 6,
  cut = 0.05,
  expansion = 2,
  inflation = 1,
  cor.fn = "cor",
  iter = 1000
)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{beta}{an integer, the soft-thresholding power for WGCNA::adjacency}

\item{cut}{a numeric between 0 and 1, the adjacency threshold used to binarize the weighted adjacency matrix}

\item{expansion}{an integer, the expansion parameter of the MCL algorithm}

\item{inflation}{an integer, the inflation parameter of the MCL algorithm}

\item{cor.fn}{a character, the correlation function to use in WGCNA::adjacency; either 'bicor' or 'cor'}

\item{iter}{an integer, the maximum number of iterations for the MCL algorithm}
}
\value{
a module object
}
\description{
Detect co-expression modules from a data matrix using Markov Clustering (MCL)
}
