% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{hybrid_modules}
\alias{hybrid_modules}
\title{Uses WGCNA and ICA to indentify modules}
\usage{
hybrid_modules(
  x,
  min.size,
  max.size,
  min.sft = 0.85,
  beta = NULL,
  cor.FN = "bicor",
  powers = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  hclust.method = "average",
  cut.height = NULL,
  core.size = 10,
  starting.threshold = 0.8,
  min.threshold = 0
)
}
\arguments{
\item{x}{a p x n  matrix of features}

\item{min.size}{minimum size of output modules}

\item{max.size}{maximum size of output modules}

\item{min.sft}{an integer, the min sft used by the pickSoftThreshold function}

\item{beta}{an integer, the power value used by WGCNA::adjacency}

\item{cor.FN}{a character, the cor.options WGCNA::adjacency}

\item{powers}{an integer vector, the powers vector used by WGCNA::pickSoftThreshold}

\item{hclust.method}{a character, the method used by flashClust}

\item{cut.height}{an integer between 0 and 1, the cut height used by cutreeDynamic}

\item{core.size}{the size of the module cores determine by ICA}

\item{starting.threshold}{the starting threshold for correlation to be added to a module}

\item{min.threshold}{the minimun threshold for correlation, if set above zero will result in incomplete assignments}
}
\value{
a module object
}
\description{
Uses WGCNA and ICA to indentify modules
}
