% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{make_lft}
\alias{make_lft}
\title{Wrapper for netutils sample_lfr that adds name and "module" label to nodes}
\usage{
make_lft(
  n = 120,
  tau1 = 3,
  tau2 = 2,
  mu = 0.08,
  average.degree = 6,
  max.degree = 10,
  min.community = 30,
  max.community = 50
)
}
\arguments{
\item{n}{a integer, the number of nodes}

\item{tau1}{PLACEHOLDER}

\item{tau2}{PLACEHOLDER}

\item{mu}{PLACEHOLDER}

\item{average.degree}{PLACEHOLDER}

\item{max.degree}{PLACEHOLDER}

\item{min.community}{PLACEHOLDER}

\item{max.community}{PLACEHOLDER}
}
\value{
an igraph object
}
\description{
Wrapper for netutils sample_lfr that adds name and "module" label to nodes
}
