% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_plot.R
\name{modular_plot}
\alias{modular_plot}
\title{Plot a modular graph with nodes colored by module membership using visNetwork}
\usage{
modular_plot(g, module.name = "module")
}
\arguments{
\item{g}{an igraph object whose nodes have a module vertex attribute}

\item{module.name}{a character, the name of the vertex attribute storing module membership (default 'module')}
}
\value{
a visNetwork HTML widget
}
\description{
Plot a modular graph with nodes colored by module membership using visNetwork
}
