% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{nodewise_fuzzy_modules}
\alias{nodewise_fuzzy_modules}
\title{Creates overlapping sets of fuzzy modules based on nodewise correlation of nodes outside the module with any of the nodes within it}
\usage{
nodewise_fuzzy_modules(x, input.modules, max.size)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{input.modules}{a module object}

\item{max.size}{The maximum number of members in a data matrix}
}
\value{
a module object
}
\description{
Creates overlapping sets of fuzzy modules based on nodewise correlation of nodes outside the module with any of the nodes within it
}
