% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{percent_module_match}
\alias{percent_module_match}
\title{Calculate the percentage of nodes assigned to the same module across two module sets, compared by node index}
\usage{
percent_module_match(m1, m2)
}
\arguments{
\item{m1}{a module S4 object containing module membership assignments}

\item{m2}{a module S4 object containing module membership assignments to compare against m1}
}
\value{
an integer between 0 and 100
}
\description{
Calculate the percentage of nodes assigned to the same module across two module sets, compared by node index
}
