% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outdated.R
\name{pragmatic_modules}
\alias{pragmatic_modules}
\title{Detect modules using ICA then subdivide any modules exceeding a maximum size}
\usage{
pragmatic_modules(x, max.size, n.mods = NULL)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{max.size}{an integer, the maximum number of nodes allowed in any single module}

\item{n.mods}{an integer, the number of independent components to extract; if NULL it is selected automatically}
}
\value{
a module object
}
\description{
Detect modules using ICA then subdivide any modules exceeding a maximum size
}
