% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_modules.R
\name{true_modules}
\alias{true_modules}
\title{Extract true module assignments from a simulated graph}
\usage{
true_modules(g)
}
\arguments{
\item{g}{an igraph object with a 'module' vertex attribute defining ground-truth module membership for each node}
}
\value{
a module object
}
\description{
Extract true module assignments from a simulated graph
}
