% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{adj_fuzzy_modules}
\alias{adj_fuzzy_modules}
\title{Expand non-overlapping modules to fuzzy (overlapping) modules using a thresholded WGCNA adjacency matrix}
\usage{
adj_fuzzy_modules(x, adj, input.modules, max.size, ratio)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{adj}{a p x p numeric adjacency matrix (e.g. from WGCNA::adjacency) defining pairwise feature similarity}

\item{input.modules}{a module S4 object containing the non-overlapping module assignments to expand}

\item{max.size}{an integer, the maximum number of nodes allowed in any fuzzy module}

\item{ratio}{a numeric, the maximum ratio of fuzzy module size to original module size}
}
\value{
a module object
}
\description{
Expand non-overlapping modules to fuzzy (overlapping) modules using a thresholded WGCNA adjacency matrix
}
