% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{adj_fuzzy_modules}
\alias{adj_fuzzy_modules}
\title{Creates overlapping sets of fuzzy modules based on thresholded adj matrix from WGCNA}
\usage{
adj_fuzzy_modules(x, adj, input.modules, max.size, ratio)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{adj}{the thresholded adjacency matrix from WGCNA}

\item{input.modules}{a module object}

\item{max.size}{the max size of fuzzy modules}

\item{ratio}{the ratio of original nodes in the module to nodes in the fuzzy module}
}
\value{
a module object
}
\description{
Creates overlapping sets of fuzzy modules based on thresholded adj matrix from WGCNA
}
