% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{calc_F1}
\alias{calc_F1}
\title{Calculates F1 score from a true and predicted graph}
\usage{
calc_F1(g.true, g.pred)
}
\arguments{
\item{g.true}{an igraph object, the true graph for the purpose of F1 calculation}

\item{g.pred}{an igraph object, the predicted graph for the purpose of F1 calculation}
}
\value{
a list of F1 score and its components
}
\description{
Calculates F1 score from a true and predicted graph
}
