% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{calc_F1}
\alias{calc_F1}
\title{Calculate the F1 score comparing a predicted graph against a true graph}
\usage{
calc_F1(g.true, g.pred)
}
\arguments{
\item{g.true}{an igraph object, the true reference graph}

\item{g.pred}{an igraph object, the predicted graph to evaluate}
}
\value{
a list of F1 score and its components
}
\description{
Calculate the F1 score comparing a predicted graph against a true graph
}
