% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b.complexICA.R
\name{complex_ICA_modules}
\alias{complex_ICA_modules}
\title{Uses ICA to find the core of modules, then iteratively expands outward based on a decreasing threshold of correlation}
\usage{
complex_ICA_modules(
  x,
  n.comp,
  core.size = 5,
  starting.threshold = 0.5,
  min.threshold = 0
)
}
\arguments{
\item{x}{PLACEHOLDER}

\item{n.comp}{PLACEHOLDER}

\item{core.size}{PLACEHOLDER}

\item{starting.threshold}{PLACEHOLDER}

\item{min.threshold}{PLACEHOLDER}
}
\value{
a module object
}
\description{
Uses ICA to find the core of modules, then iteratively expands outward based on a decreasing threshold of correlation
}
