% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outdated.R
\name{create_overlap_modules}
\alias{create_overlap_modules}
\title{Create overlapping modules by merging pairs of non-overlapping modules based on shared nodes}
\usage{
create_overlap_modules(x, input.modules, use.eigen = TRUE, best.pairs = TRUE)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{input.modules}{a module S4 object containing the non-overlapping module assignments to expand}

\item{use.eigen}{a logical, if TRUE module pairs are selected based on eigengene correlation; if FALSE nodewise correlation is used}

\item{best.pairs}{a logical, if TRUE only the best-correlated module pairs are overlapped; if FALSE all pairs are overlapped}
}
\value{
a module object
}
\description{
Create overlapping modules by merging pairs of non-overlapping modules based on shared nodes
}
