% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{create_overlap_modules}
\alias{create_overlap_modules}
\title{Create a set of overlapping modules from data matrix and strict, none overlapping modules}
\usage{
create_overlap_modules(x, input.modules, use.eigen = TRUE, best.pairs = TRUE)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{input.modules}{a module object}

\item{use.eigen}{whether or not to use the eigen gene to determine overlaps, otherwise nodewise correlation is used}

\item{best.pairs}{whether or not to overlap the best pairs of modules based on eigen genes only, otherwise all pairs of models are overlapped}
}
\value{
a module object
}
\description{
Create a set of overlapping modules from data matrix and strict, none overlapping modules
}
