% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{divide_and_conquer}
\alias{divide_and_conquer}
\title{Learns a network from data by dividing into it modules which are learned individually and stitched back together}
\usage{
divide_and_conquer(
  x,
  subgraph.module,
  keep.all.edges = FALSE,
  n.cores = 1,
  graph.learning.func = learn_SILGGM_graph,
  arg.wrapping.func = .default_arg_wrapper,
  out.parsing.func = .default_output_parser,
  packages.to.each = c("SILGGM", "igraph"),
  export.to.each = c("learn_SILGGM_graph", ".upper_tri_to_matrix", ".matrix_p_adjust",
    ".upper_tri_vec", ".pcor_zscore", ".pvalue", ".abs_pcor_filter"),
  ...
)
}
\arguments{
\item{x}{input data, p x n matrix}

\item{subgraph.module}{overlappping subsets of p wihtin a module object}

\item{keep.all.edges}{whether to keep all edges when combine subgraphs}

\item{n.cores}{number of cores to use for multithreading}

\item{graph.learning.func}{function to learn graphs}

\item{arg.wrapping.func}{function to wrap each module of x with args for graph learning function}

\item{out.parsing.func}{function to parse outputs of graph learning into list of igraphs and other outputs}

\item{packages.to.each}{packages for each thread}

\item{export.to.each}{functions for each thread}

\item{...}{other args to pass through arg wrapper to graph learning}
}
\value{
a list containing a the full graph (an igraph objects), the modular subgraphs (also igraphs), and any other outputs of graph learning
}
\description{
Learns a network from data by dividing into it modules which are learned individually and stitched back together
}
