% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{divide_and_conquer}
\alias{divide_and_conquer}
\title{Learns a network from data by dividing into it modules which are learned individually and stitched back together}
\usage{
divide_and_conquer(
  x,
  subgraph.features,
  keep.all.edges = FALSE,
  n.cores = 1,
  graph.learning.func = learn_SILGGM_graph,
  arg.wrapping.func = .default_arg_wrapper,
  out.parsing.func = .default_output_parser,
  packages.to.each = c("SILGGM", "igraph"),
  export.to.each = c("learn_SILGGM_graph", ".upper_tri_to_matrix", ".matrix_p_adjust",
    ".upper_tri_vec", ".pcor_zscore", ".pvalue", ".abs_pcor_filter"),
  ...
)
}
\arguments{
\item{x}{PLACEHOLDER}

\item{subgraph.features}{PLACEHOLDER}

\item{keep.all.edges}{PLACEHOLDER}

\item{n.cores}{number of cores to use for multithreading}

\item{graph.learning.func}{PLACEHOLDER}

\item{arg.wrapping.func}{PLACEHOLDER}

\item{out.parsing.func}{PLACEHOLDER}

\item{packages.to.each}{PLACEHOLDER}

\item{export.to.each}{PLACEHOLDER}

\item{...}{PLACEHOLDER}
}
\value{
a list containing a the full graph (an igraph objects), the modular subgraphs (also igraphs), and any other outputs of graph learning
}
\description{
Learns a network from data by dividing into it modules which are learned individually and stitched back together
}
