% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.bd_arg_wrapper}
\alias{.bd_arg_wrapper}
\title{arg wrapper for BD graph}
\usage{
.bd_arg_wrapper(sub.x)
}
\arguments{
\item{sub.x}{subset of a large data matrix x defined based on modules}
}
\value{
a list containing the learned subgraphs data transformed for bdgraph
}
\description{
arg wrapper for BD graph
}
\keyword{internal}
