% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.bd_wrapper}
\alias{.bd_wrapper}
\title{grapgh learning wrapper for BD graph}
\usage{
.bd_wrapper(x)
}
\arguments{
\item{x}{subset of a large data matrix x defined based on modules, transformed by arg wrapper}
}
\value{
a list containing the learned subgraphs data transformed
}
\description{
grapgh learning wrapper for BD graph
}
\keyword{internal}
