% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.bd_wrapper}
\alias{.bd_wrapper}
\title{Graph learning wrapper for BDgraph: fits a Gaussian graphical model and returns an igraph object}
\usage{
.bd_wrapper(x)
}
\arguments{
\item{x}{a numeric matrix with n samples (rows) and p features (columns), as prepared by .bd_arg_wrapper}
}
\value{
a list containing the learned subgraphs data transformed
}
\description{
Graph learning wrapper for BDgraph: fits a Gaussian graphical model and returns an igraph object
}
\keyword{internal}
