% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.connect_subgraphs}
\alias{.connect_subgraphs}
\title{Connects the subgraphs output based on overlapping nodes}
\usage{
.connect_subgraphs(sub.graphs, keep.all.edges = FALSE)
}
\arguments{
\item{sub.graphs}{PLACEHOLDER}

\item{keep.all.edges}{PLACEHOLDER}
}
\value{
an igraph object comprised of the combine subgraphs
}
\description{
Connects the subgraphs output based on overlapping nodes
}
\keyword{internal}
