% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.connect_subgraphs}
\alias{.connect_subgraphs}
\title{Combine a list of sub-graphs with overlapping nodes into a single graph by reconciling edge disagreements}
\usage{
.connect_subgraphs(sub.graphs, keep.all.edges = FALSE)
}
\arguments{
\item{sub.graphs}{a list of igraph objects with overlapping node names}

\item{keep.all.edges}{a logical, if TRUE all edges are retained; if FALSE edges that appear in only one of two overlapping sub-graphs are removed}
}
\value{
an igraph object comprised of the combine subgraphs
}
\description{
Combine a list of sub-graphs with overlapping nodes into a single graph by reconciling edge disagreements
}
\keyword{internal}
