% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.default_arg_wrapper}
\alias{.default_arg_wrapper}
\title{Basic arg wrapper to prepare args for graph learning in divide_and_conquer}
\usage{
.default_arg_wrapper(
  sub.x,
  method = "B_NW_SL",
  global = T,
  alpha = 0.05,
  fdr.filter = T,
  max.fdr = 0.05,
  pos.cut = 0,
  neg.cut = 0
)
}
\arguments{
\item{sub.x}{subset of a large data matrix x defined based on modules}

\item{method}{Methods for statistical inference with 5 options: "B_NW_SL", "D-S_NW_SL", "D-S_GL", "GFC_SL" and "GFC_L"}

\item{global}{If global = TRUE, the global inference of all gene pairs is performed.}

\item{alpha}{A user-supplied sequence of pre-sepecified alpha levels for FDR control. The default is alpha = 0.05, 0.1 if no sequence is provided.}

\item{fdr.filter}{Whether to filter adjacency matrix results based on}

\item{max.fdr}{The threshold to define an edge based on partial cor q val}

\item{pos.cut}{Threhold to zero values in adjacency matrix}

\item{neg.cut}{Threhold to zero values in adjacency matrix}
}
\value{
a list of arguments to feed to graph learning function
}
\description{
Basic arg wrapper to prepare args for graph learning in divide_and_conquer
}
\keyword{internal}
