% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.default_arg_wrapper}
\alias{.default_arg_wrapper}
\title{Default argument wrapper for divide_and_conquer: packages each module data subset with SILGGM arguments}
\usage{
.default_arg_wrapper(
  sub.x,
  method = "B_NW_SL",
  global = TRUE,
  alpha = 0.05,
  fdr.filter = TRUE,
  max.fdr = 0.05,
  pos.cut = 0,
  neg.cut = 0
)
}
\arguments{
\item{sub.x}{a list of numeric matrices, each being a p_i x n subset of x for one module (features as rows)}

\item{method}{a character, the statistical inference method; one of 'B_NW_SL', 'D-S_NW_SL', 'D-S_GL', 'GFC_SL', or 'GFC_L'}

\item{global}{a logical, if TRUE global inference is performed across all feature pairs}

\item{alpha}{a numeric or numeric vector, the pre-specified FDR significance level(s) for edge inclusion (default 0.05)}

\item{fdr.filter}{a logical, if TRUE the adjacency matrix is filtered to retain only edges below max.fdr}

\item{max.fdr}{a numeric, the FDR q-value threshold above which an edge is set to zero}

\item{pos.cut}{a numeric, the minimum absolute partial correlation for a positive edge to be retained}

\item{neg.cut}{a numeric, the minimum absolute partial correlation for a negative edge to be retained}
}
\value{
a list of arguments to feed to graph learning function
}
\description{
Default argument wrapper for divide_and_conquer: packages each module data subset with SILGGM arguments
}
\keyword{internal}
