% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.default_output_parser}
\alias{.default_output_parser}
\title{Processes output of graph learning function into list of learned graphs and other outputs}
\usage{
.default_output_parser(graph.learning.outputs)
}
\arguments{
\item{graph.learning.outputs}{list of outputs from each graph learning call on a subgraph}
}
\value{
a list containing the learned subgraphs and any other outputs of graph learning
}
\description{
Processes output of graph learning function into list of learned graphs and other outputs
}
\keyword{internal}
