% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.default_output_parser}
\alias{.default_output_parser}
\title{Default output parser for divide_and_conquer: returns the graph learning outputs as learned graphs with no other outputs}
\usage{
.default_output_parser(graph.learning.outputs)
}
\arguments{
\item{graph.learning.outputs}{a list of outputs, one per sub-graph, returned by graph.learning.func}
}
\value{
a list containing the learned subgraphs and any other outputs of graph learning
}
\description{
Default output parser for divide_and_conquer: returns the graph learning outputs as learned graphs with no other outputs
}
\keyword{internal}
