% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.fit_to_max_size}
\alias{.fit_to_max_size}
\title{Helper to find_WGCNA_mods that subdivides modules exceeding a maximum size}
\usage{
.fit_to_max_size(index.vector, adj, max.size, verbose = F)
}
\arguments{
\item{index.vector}{an integer vector of length p, assigning each node to a module (0 = unassigned)}

\item{adj}{a p x p numeric adjacency matrix produced by WGCNA::adjacency}

\item{max.size}{an integer, the maximum number of nodes allowed in a module}

\item{verbose}{a logical, if TRUE progress messages are printed for each split (useful for debugging)}
}
\value{
an integer vector of length p with the updated module assignments
}
\description{
Helper to find_WGCNA_mods that subdivides modules exceeding a maximum size
}
\keyword{internal}
