% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.fit_to_max_size}
\alias{.fit_to_max_size}
\title{Helper to find_WGCNA_mods that forces all modules to be within some max size}
\usage{
.fit_to_max_size(index.vector, adj, max.size)
}
\arguments{
\item{index.vector}{index vector defining a module for each node or 0 for unassigned}

\item{adj}{adjacency matrix produced by WGCNA::adjacency}

\item{max.size}{integer for max number of nodes in each module}
}
\value{
index vector with each module fitted to max size
}
\description{
Helper to find_WGCNA_mods that forces all modules to be within some max size
}
\keyword{internal}
