% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{.halfmin_impute}
\alias{.halfmin_impute}
\title{Preforms halfmin imputation to fill NAs in matrix missing values, useful for filling NAs in matrix before graph learning}
\usage{
.halfmin_impute(dat)
}
\arguments{
\item{dat}{data matrix of numeric values}
}
\value{
the data matrix with NAs filled
}
\description{
Preforms halfmin imputation to fill NAs in matrix missing values, useful for filling NAs in matrix before graph learning
}
\keyword{internal}
