% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{.halfmin_impute}
\alias{.halfmin_impute}
\title{Impute missing values in a data matrix using half the row minimum}
\usage{
.halfmin_impute(dat)
}
\arguments{
\item{dat}{a numeric matrix with p features (rows) and n samples (columns)}
}
\value{
the data matrix with NAs filled
}
\description{
Impute missing values in a data matrix using half the row minimum
}
\keyword{internal}
