% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{.label_mods}
\alias{.label_mods}
\title{Helper to lable the modules of a graph generated by make module graph}
\usage{
.label_mods(g, n.nodes, n.mods, mod.names)
}
\arguments{
\item{g}{an igraph object}

\item{n.nodes}{an integer determining the number of features/nodes in the graph}

\item{n.mods}{an integer determining the number of modules in output graph}

\item{mod.names}{a character vector of mod names}
}
\value{
an igraph object with labeled modules for each node
}
\description{
Helper to lable the modules of a graph generated by make module graph
}
\keyword{internal}
