% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c.AssessDetectedModules.R
\name{.match_modules}
\alias{.match_modules}
\title{Finds the best module matches between two sets of nodes}
\usage{
.match_modules(m1.nodes, m2.nodes, m1.names, m2.names)
}
\arguments{
\item{m1.nodes}{Node index vector from m1}

\item{m2.nodes}{Node index vector from m2}

\item{m1.names}{Node character vector from m1}

\item{m2.names}{Node character vector from m2}
}
\value{
a list of all the best modules matched based on node membership
}
\description{
Finds the best module matches between two sets of nodes
}
\keyword{internal}
