% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{.matrix_p_adjust}
\alias{.matrix_p_adjust}
\title{Perform adjustments of p-values on a p x p matrix}
\usage{
.matrix_p_adjust(mx_p)
}
\arguments{
\item{mx_p}{a p x p matrix, (i,j) represents the p-value of the partial correlation between node i and node j}
}
\value{
a p x p matrix, (i,j) represents the adjusted p-value of the partial correlation between node i and node j
}
\description{
Perform adjustments of p-values on a p x p matrix
}
\keyword{internal}
