% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.merge_modules}
\alias{.merge_modules}
\title{Helper that merges similar modules by eigengene similarity (WGCNA::mergeCloseModules)}
\usage{
.merge_modules(
  index.vector,
  t.x,
  cor.FN,
  cor.options,
  merging.cut,
  min.mods,
  max.size
)
}
\arguments{
\item{index.vector}{an integer vector of length p assigning each node to a module (0 = unassigned)}

\item{t.x}{a samples x features numeric matrix (i.e. t(x))}

\item{cor.FN}{a character, the correlation function ('bicor' or 'cor')}

\item{cor.options}{a list of correlation options passed to mergeCloseModules}

\item{merging.cut}{a numeric, the eigengene dissimilarity threshold for merging}

\item{min.mods}{an integer, the minimum number of modules to retain}

\item{max.size}{a numeric, modules are only merged while all are below this size}
}
\value{
an integer vector of length p with merged module assignments
}
\description{
Helper that merges similar modules by eigengene similarity (WGCNA::mergeCloseModules)
}
\keyword{internal}
