% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.module_check}
\alias{.module_check}
\title{Checks if a module object is properly formated and throws errors if not}
\usage{
.module_check(x, m)
}
\arguments{
\item{x}{data matrix used to learn the modules}

\item{m}{a module object}
}
\value{
TRUE, if all checks passed
}
\description{
Checks if a module object is properly formated and throws errors if not
}
\keyword{internal}
