% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.module_check}
\alias{.module_check}
\title{Validate a module object against the data matrix and throw errors if checks fail}
\usage{
.module_check(x, m)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{m}{a module S4 object containing module membership assignments}
}
\value{
TRUE, if all checks passed
}
\description{
Validate a module object against the data matrix and throw errors if checks fail
}
\keyword{internal}
