% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{.pvalue}
\alias{.pvalue}
\title{Compute the two-sided p-value of a z-score}
\usage{
.pvalue(z_score)
}
\arguments{
\item{z_score}{a z-score}
}
\value{
the p-value
}
\description{
Compute the two-sided p-value of a z-score
}
\keyword{internal}
