% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Multithreaded_DaC.R
\name{.RSNet_output_parser}
\alias{.RSNet_output_parser}
\title{Output parser for RSNet learning}
\usage{
.RSNet_output_parser(graph.learning.outputs)
}
\arguments{
\item{graph.learning.outputs}{list of outputs from each graph learning call on a subgraph}
}
\value{
a list containing the learned subgraphs and any other outputs of graph learning
}
\description{
Output parser for RSNet learning
}
\keyword{internal}
