% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{.sft_check}
\alias{.sft_check}
\title{Helper to WGCNA calling functions to select soft thresholding power in WGCNA::adjacency}
\usage{
.sft_check(sft)
}
\arguments{
\item{sft}{output of WGCNA::pickSoftThreshold}
}
\value{
an int power level
}
\description{
Helper to WGCNA calling functions to select soft thresholding power in WGCNA::adjacency
}
\keyword{internal}
