% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{eigen_fuzzy_modules}
\alias{eigen_fuzzy_modules}
\title{Creates overlapping sets of fuzzy modules based on correlation of nodes outside the module with its eigen gene}
\usage{
eigen_fuzzy_modules(x, input.modules, max.size, n.pc = 2, ratio = 1.5)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{input.modules}{a module object}

\item{max.size}{the maxsize of fuzzy modules}

\item{n.pc}{number of principle componets use to define correlation of fuzzy nodes to the module}

\item{ratio}{the ratio of original nodes in the module to nodes in the fuzzy module}
}
\value{
a module object
}
\description{
Creates overlapping sets of fuzzy modules based on correlation of nodes outside the module with its eigen gene
}
