% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{eigen_fuzzy_modules}
\alias{eigen_fuzzy_modules}
\title{Creates overlapping sets of fuzzy modules based on correlation of nodes outside the module with its eigen gene}
\usage{
eigen_fuzzy_modules(x, input.modules, max.size)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{input.modules}{a module object}

\item{max.size}{The maximum number of members in a data matrix}
}
\value{
a module object
}
\description{
Creates overlapping sets of fuzzy modules based on correlation of nodes outside the module with its eigen gene
}
