% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_ICA_mods}
\alias{find_ICA_mods}
\title{Detect co-expression modules from a data matrix using Independent Component Analysis (ICA)}
\usage{
find_ICA_mods(x, n.comp, ...)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{n.comp}{an integer, the number of independent components (modules) to extract}

\item{...}{additional arguments passed to fastICA::fastICA}
}
\value{
a module object
}
\description{
Detect co-expression modules from a data matrix using Independent Component Analysis (ICA)
}
