% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_ICA_mods}
\alias{find_ICA_mods}
\title{Uses ICA to detect modules from a data matrix}
\usage{
find_ICA_mods(x, n.comp, ...)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{n.comp}{the number of components to search for in the data}

\item{...}{other arguments to fastICA}
}
\value{
a module object
}
\description{
Uses ICA to detect modules from a data matrix
}
