% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_mcl_mods}
\alias{find_mcl_mods}
\title{Uses MCL to detect modules from a data matrix}
\usage{
find_mcl_mods(
  x,
  beta = 6,
  cut = 0.05,
  expansion = 2,
  inflation = 1,
  cor.fn = "cor",
  iter = 1000
)
}
\arguments{
\item{x}{a n x p matrix of features}

\item{beta}{an integer, the power value for WGCNA::adjacency}

\item{cut}{a numeric < 1, the cut threshold for converting from correlation to binary adjacency}

\item{expansion}{an integer, the expansion value of the mcl function}

\item{inflation}{an integer, the inflation value of the mcl function}

\item{cor.fn}{the correlation function to use in WGCNA::adjacency}

\item{iter}{an integer, the maximum number of iterations in the mcl function}
}
\value{
a module object
}
\description{
Uses MCL to detect modules from a data matrix
}
