% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_megena_mods}
\alias{find_megena_mods}
\title{Detect co-expression modules from a data matrix using MEGENA}
\usage{
find_megena_mods(
  x,
  method = "pearson",
  fdr.cutoff = 0.05,
  mod.pval = 0.05,
  hub.pval = 0.05,
  cor.perm = 10,
  hub.perm = 100,
  min.size = 20,
  is.signed = FALSE,
  n.cores = 1
)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{method}{a character, the correlation method passed to MEGENA::calculate.correlation}

\item{fdr.cutoff}{a numeric, the FDR cutoff used in MEGENA::calculate.correlation}

\item{mod.pval}{a numeric, the module significance p-value threshold for MEGENA::do.MEGENA}

\item{hub.pval}{a numeric, the hub significance p-value threshold for MEGENA::do.MEGENA}

\item{cor.perm}{an integer, the number of permutations used in MEGENA::calculate.correlation}

\item{hub.perm}{an integer, the number of permutations used for hub identification in MEGENA::do.MEGENA}

\item{min.size}{an integer, the minimum number of nodes allowed in a module}

\item{is.signed}{a logical, if TRUE signed correlation is used in MEGENA::calculate.correlation}

\item{n.cores}{an integer, the number of cores to use for parallel processing}
}
\value{
a module object
}
\description{
Detect co-expression modules from a data matrix using MEGENA
}
