% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_megena_mods}
\alias{find_megena_mods}
\title{Uses megena to detect modules from a data matrix}
\usage{
find_megena_mods(
  x,
  method = "pearson",
  fdr.cutoff = 0.05,
  mod.pval = 0.05,
  hub.pval = 0.05,
  cor.perm = 10,
  hub.perm = 100,
  min.size = 20,
  is.signed = FALSE,
  n.cores = 1
)
}
\arguments{
\item{x}{a n x p matrix of features}

\item{method}{method for the calculate.correlation function}

\item{fdr.cutoff}{fdr cut off value in calculate.correlation}

\item{mod.pval}{module p value for the do.MEGENA function}

\item{hub.pval}{hub p value for the do.MEGENA function}

\item{cor.perm}{number of  permutations in calculate.correlation}

\item{hub.perm}{the number of hub permutations in do.MEGENA}

\item{min.size}{the minimum number of nodes in each module}

\item{is.signed}{indicate using signed/unsigned correlation in calculate.correlation}

\item{n.cores}{number of cores to use for multithreading}
}
\value{
a module object
}
\description{
Uses megena to detect modules from a data matrix
}
