% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_WGCNA_mods}
\alias{find_WGCNA_mods}
\title{Uses WGCNA to detect modules from a data matrix}
\usage{
find_WGCNA_mods(
  x,
  min.size = 10,
  max.size = NULL,
  min.sft = 0.85,
  beta = NULL,
  cor.FN = c("bicor", "cor"),
  powers = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  hclust.method = "average",
  cut.height = NULL,
  merge = F,
  merging.cut = 0.2,
  iterate = T
)
}
\arguments{
\item{x}{a P x N matrix of features}

\item{min.size}{an integer, the min size of modules produced by dynamicTreeCut}

\item{max.size}{an integer, the maximum size of modules}

\item{min.sft}{an integer, the min sft used by the pickSoftThreshold function}

\item{beta}{an integer, the power value used by WGCNA::adjacency}

\item{cor.FN}{a character, the cor.options WGCNA::adjacency}

\item{powers}{an integer vector, the powers vector used by WGCNA::pickSoftThreshold}

\item{hclust.method}{a character, the method used by flashClust}

\item{cut.height}{an integer between 0 and 1, the cut height used by cutreeDynamic}

\item{merge}{a boolean, whether or not to attempt to merge modules using WGCNA::mergeCloseModules}

\item{merging.cut}{a numeric between 0 and 1, the cutheight for WGCNA::mergeCloseModules}

\item{iterate}{boolean, whether to run module detection iteratively to break up large modules}
}
\value{
a module object
}
\description{
Uses WGCNA to detect modules from a data matrix
}
