% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{find_WGCNA_mods}
\alias{find_WGCNA_mods}
\title{Detect co-expression modules from a data matrix using WGCNA}
\usage{
find_WGCNA_mods(
  x,
  min.size = 10,
  max.size = NULL,
  min.sft = 0.85,
  beta = NULL,
  cor.FN = c("bicor", "cor"),
  powers = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  hclust.method = "average",
  cut.height = NULL,
  merge = F,
  merging.cut = 0.2,
  iterate = T
)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{min.size}{an integer, the minimum number of nodes allowed in a module (passed to cutreeDynamic)}

\item{max.size}{an integer, the maximum number of nodes allowed in a module}

\item{min.sft}{a numeric between 0 and 1, the minimum R-squared threshold for soft-thresholding power selection}

\item{beta}{an integer, the soft-thresholding power for WGCNA::adjacency; if NULL it is selected automatically via pickSoftThreshold}

\item{cor.FN}{a character, the correlation function to use in WGCNA::adjacency; either 'bicor' or 'cor'}

\item{powers}{an integer vector, candidate soft-thresholding powers evaluated by WGCNA::pickSoftThreshold}

\item{hclust.method}{a character, the agglomeration method passed to flashClust}

\item{cut.height}{a numeric between 0 and 1, the dendrogram cut height used by cutreeDynamic}

\item{merge}{a logical, if TRUE adjacent modules are merged using WGCNA::mergeCloseModules}

\item{merging.cut}{a numeric between 0 and 1, the eigengene dissimilarity threshold for WGCNA::mergeCloseModules}

\item{iterate}{a logical, if TRUE module detection is run iteratively to break up any modules exceeding max.size}
}
\value{
a module object
}
\description{
Detect co-expression modules from a data matrix using WGCNA
}
