% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{learn_ARACNE_graph}
\alias{learn_ARACNE_graph}
\title{Learn a gene co-expression graph from a data matrix using the ARACNE algorithm}
\usage{
learn_ARACNE_graph(x, eps = 0, threshold = 0.05)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{eps}{a numeric, the data processing inequality threshold used by ARACNE to remove indirect edges}

\item{threshold}{a numeric, the mutual information value above which an edge is retained}
}
\value{
an igraph object
}
\description{
Learn a gene co-expression graph from a data matrix using the ARACNE algorithm
}
