% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{learn_ARACNE_graph}
\alias{learn_ARACNE_graph}
\title{Learn a gene co-expression graph from a data matrix using the ARACNE algorithm}
\usage{
learn_ARACNE_graph(
  x,
  eps = 0,
  percentile.thresold = 0.95,
  mim.threshold = NULL
)
}
\arguments{
\item{x}{a numeric matrix with p features (rows) and n samples (columns)}

\item{eps}{a numeric, the data processing inequality threshold used by ARACNE to remove indirect edges}

\item{percentile.thresold}{a numeric, the top percentile of edges by strength will be kept in the final graph}

\item{mim.threshold}{a numeric, the a fixed mutual information threshold that overrides the percentile, edges bellow this strength will be removed}
}
\value{
a named list with two elements: 'graph', the learned weighted igraph object, and 'weights', the mutual information (ARACNE) matrix
}
\description{
Learn a gene co-expression graph from a data matrix using the ARACNE algorithm
}
