% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{learn_SILGGM_graph}
\alias{learn_SILGGM_graph}
\title{Learns a graph from data using SILGGM}
\usage{
learn_SILGGM_graph(
  x,
  method = "B_NW_SL",
  global = T,
  alpha = 0.05,
  fdr.filter = T,
  max.fdr = 0.05,
  pos.cut = 0,
  neg.cut = 0,
  ...
)
}
\arguments{
\item{x}{an  n by p matrix of numeric data}

\item{method}{Methods for statistical inference with 5 options: "B_NW_SL", "D-S_NW_SL", "D-S_GL", "GFC_SL" and "GFC_L"}

\item{global}{If global = TRUE, the global inference of all gene pairs is performed.}

\item{alpha}{A user-supplied sequence of pre-sepecified alpha levels for FDR control. The default is alpha = 0.05, 0.1 if no sequence is provided.}

\item{fdr.filter}{Whether to filter adjacency matrix results based on}

\item{max.fdr}{The threshold to define an edge based on partial cor q val}

\item{pos.cut}{Threhold to zero values in adjacency matrix}

\item{neg.cut}{Threhold to zero values in adjacency matrix}

\item{...}{Additional arguments passed to SILGGM function}
}
\value{
an igraph object
}
\description{
Learns a graph from data using SILGGM
}
