% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.LearnGraphs.R
\name{learn_WGCNA_graph}
\alias{learn_WGCNA_graph}
\title{Uses WGCNA to learn a graph}
\usage{
learn_WGCNA_graph(
  x,
  min.sft = 0.85,
  beta = NULL,
  cor.FN = c("bicor", "cor"),
  powers = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  threshold = NULL
)
}
\arguments{
\item{x}{a n x p matrix of features}

\item{min.sft}{an integer, the min sft used by the pickSoftThreshold function}

\item{beta}{an integer, the power value used by WGCNA::adjacency}

\item{cor.FN}{a character, the cor.options WGCNA::adjacency}

\item{powers}{an integer vector, the powers vector used by WGCNA::pickSoftThreshold}

\item{threshold}{value to binarize adj matrix against}
}
\value{
a igraph object object
}
\description{
Uses WGCNA to learn a graph
}
