% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{make_modular_graph}
\alias{make_modular_graph}
\title{Create graph comprised of multiple modules: using either Erdős–Rényi of scale-free structures within modules}
\usage{
make_modular_graph(
  g.type = c("er", "sf"),
  n.mods = 3,
  n.nodes = 120,
  n.mod.links = 3,
  no.uncon = T,
  link.all = T,
  p.edge = 0.05,
  power = 1,
  z.appeal = 1,
  ...
)
}
\arguments{
\item{g.type}{a character determining graph type graph type, er for Erdos Renyi or sf for scale free}

\item{n.mods}{an integer determining the number of modules in output graph}

\item{n.nodes}{an integer determining the number of features/nodes in the graph}

\item{n.mod.links}{an integer determining the number of edges between each module in the graph}

\item{no.uncon}{a boolean determining if nodes with zero edges are allowed}

\item{link.all}{a boolean determining whether all modules have edges between each other}

\item{p.edge}{an integer determining the odds of two nodes within a module havin an edge}

\item{power}{an integer determining the power in scale free graph generation}

\item{z.appeal}{an integer determining the z appeal of nodes in scale fre graph generation}

\item{...}{other arguments passed to graph building functions}
}
\value{
an igraph object
}
\description{
Create graph comprised of multiple modules: using either Erdős–Rényi of scale-free structures within modules
}
