% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{make_submodular_graph}
\alias{make_submodular_graph}
\title{Create graph comprised of multiple subgraphs each of which is comprised of multiple modules, see create modular graph for details}
\usage{
make_submodular_graph(
  g.type = c("er", "sf"),
  n.mods = 3,
  n.mod.links = 3,
  n.sg = 3,
  n.sg.nodes = 120,
  n.sg.links = 1,
  no.uncon = T,
  link.all = T,
  p.edge = 0.05,
  power = 1,
  z.appeal = 1,
  ...
)
}
\arguments{
\item{g.type}{a character, graph type graph type, er for Erdos Renyi or sf for scale free}

\item{n.mods}{an integer, the number of modules in output graph}

\item{n.mod.links}{an integer, the number of edges between each module in the graph}

\item{n.sg}{an integer, the number of module subgraphs}

\item{n.sg.nodes}{an integer, the number of nodes in each subgraph}

\item{n.sg.links}{an integer, the number of edges between each pair of subgraphs}

\item{no.uncon}{a boolean, if nodes with zero edges are allowed}

\item{link.all}{a boolean, whether all modules have edges between each other}

\item{p.edge}{an integer, the odds of two nodes within a module havin an edge}

\item{power}{an integer, the power in scale free graph generation}

\item{z.appeal}{an integer, the z appeal of nodes in scale fre graph generation}

\item{...}{other arguments passed to graph building functions}
}
\value{
an igraph object
}
\description{
Create graph comprised of multiple subgraphs each of which is comprised of multiple modules, see create modular graph for details
}
