% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{modular_plot}
\alias{modular_plot}
\title{Plot a modular graph with nodes colored by module membership using visNetwork}
\usage{
modular_plot(
  g,
  module.name = "module",
  subgraphs = FALSE,
  subname = "subgraph"
)
}
\arguments{
\item{g}{an igraph object whose nodes have a module (and optionally subgraph) vertex attribute}

\item{module.name}{a character, the name of the vertex attribute storing module membership (default 'module')}

\item{subgraphs}{a logical, if TRUE nodes are colored by module and can be filtered by subgraph membership}

\item{subname}{a character, the name of the vertex attribute storing subgraph membership (default 'subgraph')}
}
\value{
a visNetwork HTML widget
}
\description{
Plot a modular graph with nodes colored by module membership using visNetwork
}
