% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{modular_plot}
\alias{modular_plot}
\title{Plots a graph with nicely coloured modules}
\usage{
modular_plot(
  g,
  module.name = "module",
  subgraphs = FALSE,
  subname = "subgraph"
)
}
\arguments{
\item{g}{an igraph object}

\item{module.name}{the name of the module value on each node}

\item{subgraphs}{boolean indicating whether to use subgraphs}

\item{subname}{the name of the subgraph value on each node}
}
\description{
Plots a graph with nicely coloured modules
}
