% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{module_contiguity}
\alias{module_contiguity}
\title{Calculate the percentage of nodes with more edges within their module then outside of it}
\usage{
module_contiguity(g, test.module)
}
\arguments{
\item{g}{an igraph object}

\item{test.module}{a module to test for contiguity, whose nodes indexs and names must match the graph}
}
\value{
an integer between 0 and 100
}
\description{
Calculate the percentage of nodes with more edges within their module then outside of it
}
