% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{percent_module_match}
\alias{percent_module_match}
\title{Calculates the percentage of nodes assigned to the same module between to module sets, based on node index, not module name}
\usage{
percent_module_match(m1, m2)
}
\arguments{
\item{m1}{a module object}

\item{m2}{another module object to compare to}
}
\value{
an integer between 0 and 100
}
\description{
Calculates the percentage of nodes assigned to the same module between to module sets, based on node index, not module name
}
