% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{pragmatic_modules}
\alias{pragmatic_modules}
\title{Uses ICA to detect modules from a data matrix then reassigned nodes to fit all modules to a max size}
\usage{
pragmatic_modules(x, max.size, n.mods = NULL)
}
\arguments{
\item{x}{a p x n matrix of features}

\item{max.size}{the maxsize of modules}

\item{n.mods}{the number of modules to search for in the data}
}
\value{
a module object
}
\description{
Uses ICA to detect modules from a data matrix then reassigned nodes to fit all modules to a max size
}
