% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{sim_graph_data}
\alias{sim_graph_data}
\title{Simulate gene expression data for n.samples samples based on the covariance structure implied by a graph}
\usage{
sim_graph_data(g, n.samples, mean.vec = NULL)
}
\arguments{
\item{g}{an igraph object defining the dependency structure between features}

\item{n.samples}{an integer, the number of samples to generate}

\item{mean.vec}{a numeric vector of length p, the mean expression value for each feature; defaults to a zero vector if NULL}
}
\value{
a p x n data matrix
}
\description{
Simulate gene expression data for n.samples samples based on the covariance structure implied by a graph
}
