% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.SimulateGraphs.R
\name{sim_graph_data}
\alias{sim_graph_data}
\title{Generates simulated data moduleing n.samples based on a network}
\usage{
sim_graph_data(g, n.samples, mean.vec = NULL)
}
\arguments{
\item{g}{an igraph object}

\item{n.samples}{the number of samples to generate}

\item{mean.vec}{a vector of sample means}
}
\value{
a p x n data matrix
}
\description{
Generates simulated data moduleing n.samples based on a network
}
