% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_fuzzy.R
\name{true_fuzzy}
\alias{true_fuzzy}
\title{Expand true modules to fuzzy (overlapping) modules based on second-order graph neighborhood}
\usage{
true_fuzzy(m, g)
}
\arguments{
\item{m}{a module S4 object containing module membership assignments}

\item{g}{an igraph object with a 'module' vertex attribute defining ground-truth module membership for each node}
}
\value{
a module object
}
\description{
Expand true modules to fuzzy (overlapping) modules based on second-order graph neighborhood
}
