% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{true_fuzzy}
\alias{true_fuzzy}
\title{Returns fuzzy moduls based on neighborhood in true graph}
\usage{
true_fuzzy(m, g)
}
\arguments{
\item{m}{a module object}

\item{g}{a igraph object with values for each node in the V(g)$module field}
}
\value{
a module object
}
\description{
Returns fuzzy moduls based on neighborhood in true graph
}
