% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.DetectModules.R
\name{true_modules}
\alias{true_modules}
\title{Returns true modules from a simulated graph with V(g)$modules}
\usage{
true_modules(g)
}
\arguments{
\item{g}{a igraph object with values for each node in the V(g)$module field}
}
\value{
a module object
}
\description{
Returns true modules from a simulated graph with V(g)$modules
}
